tests: rename test_xplot.py → test_templates.py + add deprecation coverage
tests/test_xplot.py imported plot_diverging_bar through the
deprecated `dartwork_mpl.xplot` alias. Since #41 landed, that path
emits a DeprecationWarning on attribute access; the test suite should
lead, not lag.

- `git mv tests/test_xplot.py tests/test_templates.py`.
- Switch the import to the canonical path:
    `from dartwork_mpl.templates import plot_diverging_bar`.
- Refresh the module docstring to match.
- Add a new TestXplotDeprecationAlias class with two regression
  tests: one asserts the legacy `from dartwork_mpl.xplot import ...`
  still resolves to the same callable, the other asserts `dm.xplot`
  attribute access raises DeprecationWarning. This pins the
  deprecation contract so removal cannot be accidental.

Verified:
- 5/5 tests in test_templates.py pass (3 migrated + 2 new).
- Full suite: 534 passed, 3 skipped.
- ruff check / ruff format --check / mypy clean on the file.

Closes #48
