+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - **Guardrail scope expanded to `docs/` and `examples/`**: `tests/test_domain_neutrality.py` now runs a second parametrised test `test_no_finance_terms_in_docs_and_examples` that scans every `.py` / `.md` / `.rst` / `.mplstyle` file under `docs/` and the top-level `examples/` tree for the same finance-domain vocabulary (English: revenue, profit, ebitda, earnings, fiscal, valuation, dcf; Korean: 매출, 매출액, 영업이익, 억원, 조원). Build artefacts (`_build`, `examples_gallery`, `_static`, `_templates`, `__pycache__`) are excluded. The original `test_no_finance_terms_in_shipped_source` coverage is unchanged. The full guardrail now scans 184+ files per run.
 - **`dm.helpers.labels`** — new submodule name for what used to be `dm.helpers.formatting`. Houses `format_axis_labels`, `optimize_legend`, and `add_value_labels`. Renamed to avoid the naming clash with the top-level `dartwork_mpl.formatting` module (which houses the `format_axis_*` tick formatters). The old `dm.helpers.formatting` import path still works but emits a `DeprecationWarning` pointing at the new name.
+- **`tests/test_templates.py`** — renamed from `tests/test_xplot.py` to match the canonical module name. Migrated the three existing tests to use `from dartwork_mpl.templates import plot_diverging_bar` (canonical) instead of the deprecated `from dartwork_mpl.xplot import plot_diverging_bar`, and added two regression tests that exercise the `dm.xplot` deprecation shim: one asserts the legacy import still resolves to the same callable, and the other asserts `dm.xplot` attribute access emits a `DeprecationWarning`.
 - **Gallery: six new single-plot spine / grid examples** under `docs/examples_source/01_styling_and_themes/` — `plot_spine_minimal.py`, `plot_spine_visibility.py`, `plot_spine_styling.py`, `plot_grid_customization.py`, `plot_spine_publication_styles.py`, `plot_spine_dark_theme.py`, `plot_spine_dashboard.py`. Each renders exactly one figure, per the repo's one-plot-per-file convention.
 - **Gallery: six new single-plot helpers examples** under `docs/examples_source/05_advanced_components/` — `plot_helpers_data_validation.py`, `plot_helpers_color_selection.py`, `plot_helpers_labels.py` (renamed from `plot_helpers_formatting.py` for consistency with the new `dm.helpers.labels` submodule), `plot_helpers_quality.py`, `plot_helpers_io.py`, `plot_helpers_workflow.py`. Each renders exactly one figure and covers one `dm.helpers` submodule (plus an end-to-end workflow demo), replacing the monolithic six-figure `plot_helpers_usage.py`.
